requirements.txt
fastapi>=0.111,<1.0
httpx>=0.27,<1.0
pydantic-settings>=2.2,<3.0
python-multipart>=0.0.9,<1.0
pytest>=8.2,<9.0
uvicorn[standard]>=0.30,<1.0
